Open archive and database files using default handler
There are fancy internal transition animations when loading internal
file handlers, so we test whether the system regards Amaze Filemanager
AS the default handler, and then internally just transition over to the
internal handlers. If a different application is set as the system
default for these files, then create an intent and load normally.

Functional Description;
PackageManager.resolveActivity(i,f) returns a ResolveInfo of the default or only handler, null if there are no handlers, or "something else" if there are multiple handlers. We want to handle this file if we ARE DEFAULT, or there is nothing else to handle it. If there is no default set and there are multiple handlers, it will launch a chooser with the option to set a default.