Apply design pattern ( Iterator Pattern )
1. Iterator Pattern
2. utils - DataUtils.java
3. The reasons for applying the pattern are as follows.

DataUtils.java code contains the data structure used in the application.
The Data structure is a container.
The Container must be provide methods for accessing elememts sequentially.

But, adding more and more traversal algorithms to the container gradually blurs its main responsibility, which is efficient data storage.

So, the main idea of the Iterator pattern is to extract the traversal method of a collection into a separate object called "iterator."

In DataUtils.java, there are many container used in this application. And these containers are not yet optimized.
So, if you use the Iterator Pattern, you can modify these containers easily without touching whole code.