Workaround data directory access for Android >= 11 if device is rooted
Fixes #2015.

If user is on Android 11 or above, and device is rooted, when accessing Android data directory which would be blocked access, augment the path to use /data/media as prefix and list files using root. Have Toast to notify user is being redirected.

Additionally, added filter to ls results for some devices (including AVD emulator) where toybox is used instead of busybox, to filter out entries "." and "..".

Thanks again to @danielzgtg for the workaround idea!