Clear cloud connection credentials if loading failed after database migration
Fixes #3645.

After numerous attempts to fix the problem with database migration and crash on startup, 've decided to wipe out the credentials altogether.

Users just need to re-authorize access to the Cloud plugin again, that should not be a problem; but on the other hand, hope this will never be the show-breaker again - this problem had exist throughout the 3.8 cycle for too long, this must be fixed.

My apologies to all Amaze users again for the troubles.