Remove set drawer header background and Volley dependency
We are used to allow setting the drawer's background image when we still had a Google+ community. Then we disabled it and the code was left cold and dead for quite a while.

This PR removes the dead code and Volley dependency altogether, paving the way for OkHttp as the HTTP client library to use within Amaze, as well as the blockade towards implementing WebDAV client support(#376)