Fix issue ActivityNotFoundException on Utils#openURL
The previous code had a bug that makes an ActivityNotFoundException when a user tries to open a URL on a system with any browser enabled, which is common on Oculus devices, for instance. This commit uses PackageManager to check if there is any webview enabled. It includes the following changes:

- Added PackageManager query to find a match with any enabled webview.
- Starts the activity only if there is a match.
- Logs a warning when there is no webview.
- Displays a Toast message indicating that no browser is available.
- Added some translations of the message

This fix resolves the reported bug and improves the overall user experience when opening URLs.

Fixes: #3756