English GB, AU, IN: Remove untranslated strings
There is only Color that in GB is written as Colour. All other strings are the same as in en-US. It would be better to remove them because if the source string was changed then users in the en-GB locale will see an old string. In the same time if there is no a translation for a string then anyway it will be fall back to en-US